update controller
controller ipdated
